Cryptography simplified
The Cryptography class has been further simplified. The *salt* parameter
in GenerateHash() needed numerous validations to be performed as it was
provided externally. To avoid the unnecessary code bloat of validations,
the salt is now used as an **out** parameter, the value of which can be
trapped by the calling code. Thus no validations are now required. The
GenerateRandomEntropy() method which provided the salt value is now
private resulting in avoidance of the futile conversions between string
and byte[].
Unit Tests remain.